+++ container-engine-lib/docs/changelog.md
@@ -1,7 +1,7 @@
 # TBD
 
-### Features
-* Added new `KurtosisKubernetes` labels for Kubernetes resources.
+### Changes
+* Updated Kurtosis Kubernetes `label_key_consts` and tests.
 
 # 0.20.1
 ### Features